requirements.txt
absl-py==2.3.1
alembic==1.16.4
astunparse==1.6.3
axial_positional_embedding==0.3.12
blessed==1.21.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cloudpickle==3.1.1
cmaes==0.12.0
colorlog==6.9.0
CoLT5-attention==0.11.1
contourpy==1.3.2
cryptography==45.0.6
cycler==0.12.1
einops==0.8.1
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.1
fsspec==2025.7.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
gpustat==1.1.1
GPUtil==1.4.0
greenlet==3.2.4
grpcio==1.74.0
h5py==3.14.0
hyper-connections==0.2.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
lightning-utilities==0.15.2
llvmlite==0.44.0
local-attention==1.11.2
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
namex==0.1.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.65
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opt_einsum==3.4.0
optree==0.17.0
optuna==4.5.0
optunahub==0.3.1
packaging==25.0
pandas==2.3.2
patsy==1.0.1
pillow==11.3.0
product_key_memory==0.2.11
protobuf==6.32.0
psutil==7.0.0
pycparser==2.22
PyGithub==2.7.0
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
reformer-pytorch==1.4.4
requests==2.32.5
rich==14.1.0
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
shap==0.48.0
six==1.17.0
slicer==0.0.8
smmap==5.0.2
SQLAlchemy==2.0.43
statsmodels==0.14.5
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
tomli==2.2.1
torch==2.8.0
torchmetrics==1.8.1
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.3
